src/SampleSolution/Sample.Core.Tests/FizzBuzzTest.cs
using Xunit;

namespace Sample.Core.Tests
{
    public class FizzBuzzTest
    {
        [Theory]
        [InlineData(0, "0")]
        [InlineData(1, "1")]
        [InlineData(2, "2")]
        [InlineData(3, "Fizz")]
        [InlineData(4, "4")]
        [InlineData(5, "Buzz")]
        [InlineData(6, "Fizz")]
        [InlineData(7, "7")]
        [InlineData(8, "8")]
        [InlineData(9, "Fizz")]
        [InlineData(10, "Buzz")]
        [InlineData(11, "11")]
        [InlineData(12, "Fizz")]
        [InlineData(13, "13")]
        [InlineData(14, "14")]
        [InlineData(15, "FizzBuzz")]
        public void ConvertTest(int n, string result)
        {
            var fizzbuzz = new FizzBuzz();
            Assert.Equal(result, fizzbuzz.Convert(n));
        }
    }
}
